requirements.txt
python-dotenv==1.0.0
speechrecognition==3.10.0
Pillow==10.1.0
requests==2.31.0
urllib3==2.0.7   
certifi==2023.7.22   
charset-normalizer==3.3.2  
idna==3.4            
typing-extensions==4.8.0  
